+++ .env
@@ -1,0 +1,1 @@
+VITE_LOCAL_API=http://localhost:8080
