+++ package.json
@@ -15,7 +15,9 @@
     "prepare": "husky install"
   },
   "dependencies": {
+    "axios": "^1.4.0",
     "date-fns": "^2.29.3",
+    "dotenv": "^16.0.3",
     "leaflet": "^1.9.3",
     "react": "^18.2.0",
     "react-datepicker": "^4.11.0",
