+++ src/mocks/SensorData.ts
@@ -1,0 +1,10 @@
+export interface SensorData {
+  date: Date;
+  data: AirImpurityData[];
+  code: string;
+}
+
+export interface AirImpurityData {
+  name: string;
+  value: number;
+}
