+++ src/mocks/SensorLocation.ts
@@ -1,0 +1,5 @@
+export interface SensorLocation {
+  code: string;
+  n: number;
+  e: number;
+}
