+++ src/mocks/example_data/example_sensor_data.ts
@@ -1,4 +1,4 @@
-import { type SensorData } from '@/types/models/SensorData';
+import { type SensorData } from '@/mocks/SensorData';
 
 export const data: [SensorData] = [
   {
