+++ src/services/StationsService.ts
@@ -1,15 +1,12 @@
-const locationsURI = 'http://localhost:8080/locations';
+import axios from 'axios';
+
+const locationsURI = `${import.meta.env.VITE_LOCAL_API as string}/locations`;
 
 export async function getLocations() {
-  const res = await fetch(locationsURI);
-  if (res.ok) {
-    const locations = await res.json();
-    return locations;
+  const res = await axios.get(locationsURI);
+  if (res.status) {
+    return res.data;
   }
   console.log('Failed to fetch locations.');
   return [];
-}
-
-export function getLocation(sensorCode, locations) {
-  return locations.find(val => val.code === sensorCode);
 }
